List #update 'scope' method

+++ app/controllers/api/lists_controller.rb
@@ -1,6 +1,5 @@
 class Api::ListsController < ApiController # rubocop:disable Style/ClassAndModuleChildren
   before_action :authenticated?, unless: :keyed_open
-
   def index
     user = get_key_user
     lists = List.visible_to(user)
@@ -19,20 +18,29 @@ class Api::ListsController < ApiController # rubocop:disable Style/ClassAndModul
   end
 
   def update
-    list = List.find(params[:id])
-    if list.update_attributes(list_params)
-      render json: list
+    list = List.own(params[:id], get_key_user)
+    if list
+      if list.update_attributes(list_params)
+        render json: list
+      else
+        render json: { errors: list.errors.full_messages }, status: :unprocessable_entity
+      end
     else
-      render json: { errors: list.errors.full_messages }, status: :unprocessable_entity
+      render json: {}, status: :not_found
     end
   end
 
   def destroy
     list = List.find(params[:id])
-    list.destroy
-    render json: {}, status: :no_content
-  rescue ActiveRecord::RecordNotFound
-    render json: {}, status: :not_found
+    key_owner = get_key_user
+    record_owner = User.find(list.user.id)
+    if key_owner.id == record_owner.id
+      list.destroy
+      render json: {}, status: :no_content
+    else
+      ActiveRecord::RecordNotFound
+      render json: {}, status: :not_found
+    end
   end
 
   private
